refactor(progress): split into progress.rs (message) + progress_eframe.rs (renderer)
progress.rs contains only ProgressMessage — no deps beyond route.rs.
progress_eframe.rs contains ProgressPlot and all eframe/egui rendering;
gated via #[cfg(feature = "gui")] in mod.rs, so the file itself is
clean with no cfg noise.

Adds a slot for future backends (progress_tui.rs, progress_web.rs, etc.)
by the same pattern: implement against ProgressMessage, register in mod.rs.
